Adapt to change in Chrome's behavior
Chrome used to print the header row of a table only once,
even when the table spanned multiple pages.  Now it prints
it at the top of each page, so we no longer need a hack
for the non-collated version.  We still need to use the
existing workaround for the collated version, since it
uses multiple HTML table elements.